shared/pipeline/requirements.txt
google-genai>=1.0.0
pyyaml>=6.0
Pillow>=10.0
rich>=13.0
